Fix some CNTK float32 warnings

+++ MDPAdapterSensor.py
@@ -97,7 +97,7 @@ class MDPAdapterSensor(MDP):
 		# not free
 		# entries are according to states, y for columns and x for rows
 		grid_data = env.grid_data
-		walls = np.zeros((self._height,self._width))
+		walls = np.zeros((self._height,self._width), dtype=np.float32)
 		for x in range(self._width):
 			for y in range(self._height):
 				temp_wall = grid_data[x*cell_size:(x+1)*cell_size-1,y*cell_size:(y+1)*cell_size-1]
@@ -214,7 +214,7 @@ class MDPAdapterSensor(MDP):
 			feature[0] = x
 			feature[1] = y
 			"""
-			feature = np.zeros(7)
+			feature = np.zeros(7, dtype=np.float32)
 			
 			if walls[y,x] == 1:
 				feature[6] = 1
